requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
sse-starlette==2.1.3
python-docx==1.1.2
odfpy==1.4.1
anthropic==0.34.2
google-generativeai==0.8.3
gitpython==3.1.43
aiofiles==24.1.0
groq==0.11.0
httpx==0.27.2
